WIP: Allow the isAnagram function to accept an array
If the first argument is an array of strings, check if those are
anagrams of each other.
If two arguments are passed, both as strings, compare them and
check if they're anagrams.
Otherwise, throw an error about invalid argument types.

This is still a work-in-progress since I'm not validating if the
values received in the array are strings.